add a file used to clean raw data, and get better data structure.

+++ clean_data_lianjia.py
@@ -20,8 +20,8 @@ def clean_raw_data(file_name):
     basic_information = '北京' + raw_data['strict'] +raw_data['xiaoqu'] + raw_data['more_details']
     basic_information = basic_information.str.split('/', expand=True)
     clean_data['full_address'] = basic_information[0]
-    clean_data['sleeping_room'] = basic_information[1].str[0].astype(int)
-    clean_data['drawing_room'] = basic_information[1].str[2].astype(int)
+    clean_data['sleeping_room'] = basic_information[1].str[0]
+    clean_data['drawing_room'] = basic_information[1].str[2]
     clean_data['building_area'] = basic_information[2].str[:-2].astype(float)
     clean_data['direction'] = basic_information[3]
     clean_data['house_design'] = basic_information[4]
